+++ src/aui-tree/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-1600](https://issues.liferay.com/browse/AUI-1600) Tree node items should not be rearrange-able when dragging on white space
 
 ## [3.0.0pr2](https://github.com/liferay/alloy-ui/releases/tag/3.0.0pr2)