+++ src/aui-tree/HISTORY.md
@@ -4,6 +4,10 @@
 
 ## @VERSION@
 
+No registries yet.
+
+## [3.0.1](https://github.com/liferay/alloy-ui/releases/tag/3.0.1)
+
 * [AUI-835](https://issues.liferay.com/browse/AUI-835) Alloy Tree can not be used without the mouse
 * [AUI-1115](https://issues.liferay.com/browse/AUI-1115) Unable to catch the expand event on TreeView